+++ users/src/main/java/com/marketplace/users/repositories/BuyerRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BuyerRepo extends JpaRepository<BuyerEntity, Integer> {
+public interface BuyerRepo extends JpaRepository<BuyerEntity, Long> {
 }
