+++ users/src/main/java/com/marketplace/users/repositories/SellerRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SellerRepo extends JpaRepository <SellerEntity, Integer> {
+public interface SellerRepo extends JpaRepository <SellerEntity, Long> {
 }
