+++ users/src/main/java/com/marketplace/users/services/implementations/UserImpl.java
@@ -52,7 +52,7 @@ public class UserImpl implements UserService {
     }
 
     @Override
-    public UserEntity getById(int id) throws NotFoundEntityException {
+    public UserEntity getById(long id) throws NotFoundEntityException {
         if(!userRepo.findById(id).isPresent())
             throw new NotFoundEntityException("iD invalid", "", "");
         return userRepo.findById(id).get();
@@ -66,7 +66,7 @@ public class UserImpl implements UserService {
     }
 
     @Override
-    public void delete(int userId) throws NotFoundEntityException {
+    public void delete(long userId) throws NotFoundEntityException {
         if(!userRepo.findById(userId).isPresent())
             throw new NotFoundEntityException("Aucun objet","Aucun objet n'a encore été créé dans la base de donnée", localisation+"delete");
         UserEntity deleteBuyer = userRepo.getOne(userId);
